[CP-20682] xapi_gpumon: update compatibility checking assertion
We need to use the encoded_`pgpu_metadata` from the Db in place of the
pgpu reference: this allows checking the compatibility against PGPUs from
other pools. In addition we need to add the destination host to the
parameters for the error messages metadata.

Also we fix a spelling typo in the original function name.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>